test(ARC-878): add missing checkers tests and fix chain completion validator
- Add test: no legal moves = win condition
- Add test: draw by insufficient material (king vs king)
- Add test: flying king multi-jump capture chain
- Add test: incomplete capture chain rejection (single step when more captures available)
- Fix: chain completion check now runs for all capture chains, not just multi-step
- Remove dead code: getWinner() from utils (had ID-vs-color bug, never imported)
- Remove dead code: forcedCaptureChain field from CheckersState (never used)
- Split variant/edge-case tests into checkers-variants.spec.ts to stay under 500-line limit
