+++ apps/be/src/games/engines/checkers/checkers.utils.ts
@@ -440,43 +440,3 @@ export function hasAnyMoves(
     ).length > 0
   );
 }
-
-export function getWinner(
-  board: Board,
-  currentPlayerId: string,
-  currentColor: string,
-  backwardCaptures = false,
-  flyingKings = false,
-): string | null {
-  const opponentId = currentPlayerId === 'light' ? 'dark' : 'light';
-  const opponentColor = currentPlayerId === 'light' ? 'dark' : 'light';
-
-  const currentPieces = countPieces(board, currentPlayerId);
-  const opponentPieces = countPieces(board, opponentId);
-
-  if (currentPieces === 0) return opponentId;
-  if (opponentPieces === 0) return currentPlayerId;
-
-  if (
-    !hasAnyMoves(
-      board,
-      opponentId,
-      opponentColor,
-      backwardCaptures,
-      flyingKings,
-    )
-  )
-    return currentPlayerId;
-  if (
-    !hasAnyMoves(
-      board,
-      currentPlayerId,
-      currentColor,
-      backwardCaptures,
-      flyingKings,
-    )
-  )
-    return opponentId;
-
-  return null;
-}
